backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
python-multipart
passlib[bcrypt]
python-jose[cryptography]
